[REFACTOR] MealCuration home-page로 이동
